tests using headless PPSSPP

+++ test_filter.py
@@ -1,19 +1,26 @@
 
 import contextlib
 import subprocess
+import glob
 import time
 import os
 import shutil
+import sys
 import pathlib
 from typing import Iterable, Set
 
-from PIL import Image
+import pytest
+
+#from PIL import Image
 from xvfbwrapper import Xvfb
 
 PSP_DIR = pathlib.Path.home() / ".config" / "ppsspp"
 PSP_SCREEN_READY_FILE = PSP_DIR / "screen_ready.txt"
 PSP_SCREENSHOTS_DIR = PSP_DIR / "PSP" / "SCREENSHOT"
 CORRECT_SCREENSHOTS_DIR = pathlib.Path(__file__).parent / "screenshots"
+
+FAILURE_SCREENSHOT = pathlib.Path("__testfailure.bmp")
+SCREENSHOT_WE_DONT_CARE_ABOUT = pathlib.Path("__testcompare.png")
 
 def _wait_for_screen_ready_file(max_wait_time_seconds: float = 30.0, test_interval_seconds: float = 0.05) -> None:
     time_begin = time.time()
@@ -23,6 +30,9 @@ def _wait_for_screen_ready_file(max_wait_time_seconds: float = 30.0, test_interv
         time.sleep(test_interval_seconds)
     
     raise RuntimeError(f"Screen ready file not found after {max_wait_time_seconds} seconds")
+
+
+
 
 
 @contextlib.contextmanager
@@ -51,7 +61,7 @@ def _wait_for_screenshot_file_to_be_valid(screenshot_path: pathlib.Path, max_wai
     time_begin = time.time()
     while time.time() - time_begin < max_wait_time_seconds:
         try:
-            Image.open(screenshot_path)
+            return#Image.open(screenshot_path)
         except OSError:
             # File is probably still being generated
             pass
@@ -71,35 +81,34 @@ def _wait_for_new_screenshot(existing_screenshots: Set[pathlib.Path], max_wait_t
     return screenshot_path
 
 
-def _run_prx_get_screenshot(prx: str) -> pathlib.Path:
+def _run_prx_and_compare_to_screenshot(prx: pathlib.Path, compare_to_screenshot: pathlib.Path, move_screenshots_to: pathlib.Path):
+    assert not FAILURE_SCREENSHOT.exists()
+    assert not SCREENSHOT_WE_DONT_CARE_ABOUT.exists()
     if PSP_SCREEN_READY_FILE.exists():
         PSP_SCREEN_READY_FILE.unlink()
-    existing_screenshots = _get_screenshots()
 
-    # os.environ["QT_QPA_PLATFORM"] = "xcb"
-    os.environ["QT_DEBUG_PLUGINS"] = "1"
-    with Xvfb(width=1920, height=1080):
-        with _run_and_kill(["PPSSPPQt", prx]):
+    assert compare_to_screenshot.exists()
 
-            _wait_for_screen_ready_file()
-            subprocess.run(["xdotool", "search", "PPSSPP ", "windowfocus", "key", "F12"])
-            new_screenshot = _wait_for_new_screenshot(existing_screenshots)
-            return new_screenshot
+    subprocess.run(("PPSSPPHeadless", prx, f"--screenshot={compare_to_screenshot}"), check=True)    
+
+    if SCREENSHOT_WE_DONT_CARE_ABOUT.exists():
+        shutil.move(
+            SCREENSHOT_WE_DONT_CARE_ABOUT,
+            move_screenshots_to
+        )
+
+    if FAILURE_SCREENSHOT.exists():
+        shutil.move(
+            FAILURE_SCREENSHOT,
+            move_screenshots_to
+        )
+        raise RuntimeError(f"Screenshot mismatch! {compare_to_screenshot} does not match {compare_to_screenshot}")
+
+def test_against_screenshot(prx_file: str, tmp_path: pathlib.Path):
+    correct_screenshot_filename = f"{os.path.basename(prx_file)}.bmp"
+    _run_prx_and_compare_to_screenshot(prx_file, CORRECT_SCREENSHOTS_DIR / correct_screenshot_filename, tmp_path)
 
 
-def _run_prx_and_check_screenshot(prx_file: str, tmp_path: pathlib.Path, expected_screenshot_path: pathlib.Path) -> None:
-    assert expected_screenshot_path.exists()
-    psp_screenshot_path = shutil.move(
-        _run_prx_get_screenshot(prx_file),
-        tmp_path
-    )
-    CORRECT_SCREENSHOTS_DIR
-
-    screenshot_from_psp = Image.open(psp_screenshot_path)
-    correct_screenshot = Image.open(expected_screenshot_path)
-
-    # Might make some incorrect assumptions about how the data is encoded...
-    assert list(screenshot_from_psp.getdata()) == list(correct_screenshot.getdata()), f"{psp_screenshot_path} does not match {expected_screenshot_path}"
-
-def test_white_filtered_8888(tmp_path: pathlib.Path):
-    _run_prx_and_check_screenshot("build/tests/WhiteTest8888Filtered.prx", tmp_path, CORRECT_SCREENSHOTS_DIR / "white_filtered_8888.jpg")
+def test_against_screenshot_sanity_check(prx_file: str, tmp_path: pathlib.Path):
+    with pytest.raises(RuntimeError):
+        _run_prx_and_compare_to_screenshot(prx_file, CORRECT_SCREENSHOTS_DIR / "failure_sanity_check.bmp", tmp_path)
